pyproject.toml
[tool.poetry]
name = "dionysus"
version = "0.0.0"
description = "Collects data for analysing performance art data on social media"
authors = ["Yifei Yu <yyu.mam2020@london.edu>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "dionysus", from = "src" }
]
include = [
    { path = "src/dionysus/conf_default", format = "sdist" }
]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.21.0"
orjson = "^3.8.1"
dacite = "^1.6.0"
networkx = "^2.8.8"
numpy = "^1.23.4"
ipykernel = "^6.17.1"
playwright = "^1.27.1"

[tool.poetry.group.dev.dependencies]
isort = "^5.10.1"
pytest = "^7.2.0"
mypy = "^0.990"
flake8 = "^5.0.4"
coverage = {extras = ["toml"], version = "^6.5.0"}
black = {extras = ["jupyter"], version = "^22.10.0"}
yamllint = "^1.28.0"
semgrep = "^0.121.2"
poetryup = "^0.12.3"
types-requests = "^2.28.11.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
force-exclude = '''
.*/setup\.py$
'''

[tool.isort]
profile = "black"
line_length = 88
src_paths = ["src", "tests"]
extend_skip = ["setup.py"]

[tool.flake8]
extend-ignore = "E203"
exclude = [".git", "__pycache__", "build", "dist", ".venv"]
max-complexity = 10
max-line-length = 88

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "DEBUG"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
addopts = "--junitxml=test_report.xml"
testpaths = "tests"

[tool.mypy]
files = "src"
mypy_path = "src"
exclude = [
    ".venv"
]
namespace_packages = true
explicit_package_bases = true
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "networkx.*"
]
ignore_missing_imports = true
warn_return_any = false

[tool.coverage.paths]
source = ["src", "*/site-packages"]

[tool.coverage.run]
branch = true
source = ["src"]

[tool.coverage.report]
show_missing = true
skip_covered = true
fail_under = 50
exclude_lines = [
    # Have to re-enable the standard pragma
    "pragma: no cover",

    # Don't complain about missing debug-only code:
    "def __repr__",
    '''if self\.debug''',

    # Don't complain if tests don't hit defensive assertion code:
    "raise AssertionError",
    "raise NotImplementedError",

    # Don't complain if non-runnable code isn't run:
    "if 0:",
    "if __name__ == .__main__.:",

    # Don't complain about abstract methods, they aren't run:
    '''@(abc\.)?abstractmethod''',
]

[tool.coverage.html]
directory = "htmlcov"

[tool.coverage.xml]
output = "coverage.xml"
